✨(api) use jwt to create tokens
Instead of using rest_frameword_simplejwt, use jwt module.
We don't have the advantages of implemented things in simplejwt
but makes it easier for devs to manipulate how jwt tokens are handled.
An access token is created,
either with default values if user is not authenticated
or with the values of the user.
Multiple variables were renamed so that they make more sense
Every user is moderator for now, so as not to lose moderation tools.
When booked rooms will be added:
- Everyone will be a moderator on normal rooms
- only the one who booked will be moderator on booked rooms.

✅(tests): write tests for guest and logged in users

The test checks if we redirect to correct url, with token

📝(readme): update readme with new api functionalities

Also explain how to link to existing jitsi instance