🎉(root) create the project structure
Create a 4-part structure, (frontend + backend) * (lib + demo),
implementing the following tools and features:

- We can start the demo backend
- The magnify folder is a package
- It export a default set of settings and a single app "rooms"
- We can work on the rooms app and it updates on the demo backend instantly
- We can test rooms by writing tests in test/apps/rooms
- We can write react components in the folder frontend
- We can see them in isolation in a storybook
- We can test them using testing libraries
- This frontend library is imported from the frontend demo,
also up and running as "@jitsi-magnify/frontend-lib".
This is made possible by the package json in the src,
defining the workspace with the name "jitsi-magnify", and the sub projects.

This should solve issue #6 and unlock future development