♻️(project) refactor to introduce frontend and build on libraries
We want to add a react frontend and be able to publish the backend and
frontend building blocks as libraries in respectively pypi and npm.

What we want is very close to what is done in FUN's Richie project:
https://github.com/openfun/richie
So we decided to port richie's scaffolding directly to Magnify.

Then we wanted the frontend components to be distributed as a library,
so we used yarn workspaces to split the frontend in two projects:
the library called "magnify" and the site called "demo".

Written at 4 hands with Léo-Nils Boissier <leo.nils2000@gmail.com>