✨(frontlib) create component to display a single group
A group is a set of users that are allowed to do some
things in the same way. This component is a way to present
a single group. As we want to easily see users from the
group, I added avatars, but as it can be very uggly when
screens are small, special attention was given to responsiveness,
by changing the number of avatar displayed according to
the size of the screen