✨(frontlib) add component to display a list of groups
The groups are made by a users. We need a component to
displays all groups of a user, and to perform some actions.
Right now, the api is not available, so actions are not
binded to actual requests, so components are only presentational
We just need to have components to begin to work on the views
using them.