✨(frontend) create an abstraction to fetch the api
Until now, the frontend was only presentational. The obvious way
to connect to the api is to fetch things directly into the components,
using axios for instance. But this is difficult to mock and test.
Here I tried to provide some classes that represent an interface with the
api. I called them "controller". I can have a controller that
only log actions, another that just give mocks, and another that actully
perform requests on the api.