✨(frontend) add a demonstrator of fetching data using controller
I want to make it straight forward to mutate and fetch content,
so I added a demonstration of how I think we can do it using
the controller, including how to mock it for the storybook,
and for the test. To me the concern is to create something where
there is no much to think of when we want to add a route. There are
3 functions to implements, you have a few mocks to write, and that's
it. Badly, we do not use the same language between backend and
frontend, so we can't reuse the factories and the interfaces
between the two...