🎨(frontend) separate sidebar responsabilities
Sidebar had 3 responsabilities:
- Display buttons from props
- Chose which buttons are active
- Provide a layout to build the website inside it
As they are 3 completely different things,
the layout is now a different component,
and the button is in charge of deciding if it is active.
This will also help further modifications like translation.
Change the structure of the sidebar props:
Now list of lists to delimit zones that have extra gaps