✨(frontlib) create component to display and change avatar
Avatar are a cool way to identify someone, so we want a way
for the user to upload his coolest portrait. File uploader
are usually ugly, so this components aims to give directly
to the user a preview of his avatar as soon as he load it,
and we don't see any native button.