✨(frontlib) create a row to display a single meeting
This was a hard one. The meeting is on the backend a patern
to generate a lot of meetings, like "every monday, wednesday,
between X and Y, starting at time Z, for duration W". We
needed to convert that to "your next meeting is at ...", and
there is too much cases! I choose to handle these in some logic
files into the component folder, but not into the component file
itself, so the component is still pretty light.