🤡(frontlib) add the auth and user routes to the controllers
This is a first step to integrate the backend. I added the
methods to the interface. Right now, the backend is not ready yet
but we are ok on what it will be, so we can mock it and connect
the mock to the app. As described before, we will have a default
controller (connecting the backend) that currently raise "not implemented"
errors. Then we have the mock controller, for tests, and a "log"
controller, that allow to see what happends in the console, and
resolve/reject mocked values. Until now, there was only one
resolve value possible for each route, but we want to test more,
so I added a class to easily define on which args the controller
resolve/reject.