✨(frontlib) create a blinking row for loading state
As a user, I want to know that a list is loading.
I think this simple blinking row can be used in any row list
if the state is loading. I plan to use this immediatly for the
rooms list at least. This is pure css animation, so testing
can't be very much more that rendering the component