♻️(models) refactor users and groups relations with rooms/meetings
We want to be able to associate groups and users to rooms/meetings
We also need to be able to make a room/meeting public.

At first, we thought of considering that a room/meeting was public
if it did not have any users/groups associated to it but it was not
explicit enough and it was pausing a problem because public meetings
also have administrators.