✨(frontlib) create the component to embed jitsi
This component is one of the most important piece of magnify. Adding it
made me update a few methods in the controllers, factories, and add 2
new components. Moreover, this is right now mostly a proof of concept.
The component is not perfect, but it allow to connect to a meeting using
a token provided by the controller. It shows the usecase of choosing between
several meetings when available, and finaly it shows how to create a
panel to embed magnify data along to jitsi iframe. The iframe beeing
notoriously difficult to test, I couldn't add tests on this part. Moreover,
the jitsi react sdk lib is not transpiled, making it even more difficult to test