✨(api) allow getting rooms by their slug
What the front has when a user visits a room's detail page is the
slug, not the id. We want to keep the ID as primary key to the
model for coherence with other models and simplicity. We added a
check to make sure someone doesn't try to book a room with a slug
that looks like a UID which would create confusion.