♻️(models) change way to declare group administrators
We were using the same many-to-many relationship to declare members
and administrators of a group (with a parameter "is_administrator"
on the through model).

Most often, administrators are not members of a group so this
modelization was problematic. In this commit, we introduce 2 separate
many-to-many fields for members and administrators.