♻️(api) remove room token view and accept unregistered rooms on API
The existing room token view was delivering Jitsi accesses even on
rooms that are now reserved in Magnify. We only want to allow
a room after checking if it is reserved. We decided to add this
feature on the existing room API detail endpoint and handle
unregistered rooms on a 404 error when no existing room is found.