🔒️(api) hide room configuration from non admin users
The Jitsi configuration of the room was included in the payload for
all users allowed to join the room. The problem is that we have now
added the password in this configuration which should not be visible
to non-admin users. The Jitsi configuration is only necessary for
admin users who start the conference, so we can remove it from the
payload for non-admin users.

It's easier to keep it as a normal field in the serializer and remove
it explicitly for non-admin users because we need it to be updatable
for admin users.