♻️(rooms) refactor room access to RBAC and open specific API endpoints
Room user and group access rights were served on API endpoint nested under
the room API endpoint. Access control was not properly enforced and this
pattern made things fragile and unnecessarily complicated.

We open separate endpoints for room accesses and realized the necessity
to switch to role based access control in order to differentiate
administrators from owners, the latter being the only ones to be allowed
to delete a room.