♻️(meetings) make meetings timezone and DST aware
We built meetings around 2 separate date and time fields. This was
a naive approach that does not work in an international environment
with timeones and daily saving times.

We refactor meetings to fully support timezones and DST and update
tests to make sure it now works as intended. We decided to store
datetimes in UTC and add a timezone field to record the timezone of
the meeting organizer. This is the timezone in which weekdays are
derived. The day of meeting occurences may vary for other participants
depending on DST in their own timezone. Conversion between UTC and each
user's timezone is left to the frontend so that our API always talks UTC.