♻️(access) refactor to join user and group accesses in a generic object
On the frontend, we want to display user and group accesses in the same
list so a simplification is to use the same model to record them.
Also, we will need to handle accesses on other resources than rooms,
so we split the room model in a resource base model and a room model
that inherits from this model.

Meeting accesses don't require roles (a meeting will only ever belong
to a user), so we make it a separate MeetingAccess model.