✨(api) add API endpoint to serve frontend configuration
We want to avoid having to rebuild the frontend in order to run
Magnify in its default version. The idea is to let the backend
serve the configuration that the frontend needs on a simple API
endpoint. Each setting in this configuration can be modified via
environment variables in the Docker container.