👷(project) rename docker-compose production stack to demo
"production" is a misleading name for what is a demo "production"
stack in docker-compose. We don't recommend running this in production
so it's better to rename it to "demo".

Also the demo stack should run on the "Production" settings. It was
running on the "ContinuousIntegration" settings and was broken following
the latest changes to the "Test" settings from which it depends.