✨(users) allow differentiating device users from human users
Some users are in fact connected to devices in a room. We want to
be able to identify and differentiate these special users in the
code. The idea is to identify them by the fact that they should use
a different client in Keycloak and therefore have a different audience
set in their token.