🐛(core) base room.jitsi_name only on room.slug
Currently, the room.jitsi_name bound into the jwt token generated to be
authenticated on the jitsi instance is not the same as the real jitsi room the
user tries to access. It appears this difference can prevent a user to enter
into a jitsi meeting according the jitsi server configuration. As the slug of a
room is unique, we can safely use only this field to build the room.jitsi_name.