📝(frontend) update README.md
Previously, it was challenging to contribute to the project easily, especially
when it came to understanding the frontend part. In addition to the architecture
changes made in a previous pull request, we needed to add a new section to the
README.md file to describe how to launch and contribute to the frontend
project.