🐛(frontend/core) use `config.LANGUAGE_CODE` to set default locale
The default `LANGUAGE_CODE` to use in the frontend application is bound to
the config information retrieved through the backend api. Currently, this
information was not use to set the default locale of the TranslationProvider so
English was always used until user logged in and the magnify-locale is set into
the local storage.

Fix #235